move: 선착순 퀴즈 관련 컴포넌트
