--- packages/user/src/components/shared/modal/FCFSModal.tsx
@@ -1,5 +1,0 @@
-import Modal, { ModalProps } from 'src/components/common/Modal.tsx';
-
-export default function FCFSModal(props: ModalProps) {
-	return <Modal {...props}>FCFSModal</Modal>;
-}
